[PATCH] change inputs and move header an FilterMenu
