+++ src/components/Input/Input.css
@@ -1,4 +1,8 @@
-
+.input_wrapper{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+}
 .icons {
     font-size: 1rem;
     color: var(--primaryBorderColor);
@@ -8,7 +12,7 @@
 
 .input_container {
     display: flex;
-    align-items: center;
+    width: 100%;
 }
 .select_label{
     font-size: 1rem;
@@ -21,12 +25,15 @@
     box-shadow: none;
     outline: none;
     background: transparent;
-    color:var(--primaryTextColor);
+    color:var(--primaryTextColor); 
     font-size: 1.2rem;
     border-bottom:  1px solid var(--secondaryTextColor);
+    margin-bottom: 10px;
+    width: 80%;
 }
 
 .input_container input::placeholder {
     font-weight: lighter;
     color: var(--primaryBorderColor);
 }
+
