+++ src/layout/Header/Header.css
@@ -7,7 +7,8 @@
     flex-flow: wrap;
     padding: 10px;
     gap: 20px;
-    top:0;
+    left: 0;
+    top: 0;
 }
 
 .header .header_logo_container{
